Create .exe file with pyinstaller
py2exe is unuseful;
by the way,rename MyDownloader.py
