+++ .gitignore
@@ -1,2 +1,4 @@
 *.pyc
 setting
+build/*
+dist/*
